Refactor Duke to Homie
Rename all classes, methods and variables related to duke to homie.
