--- src/main/java/duke/DukeException.java
@@ -1,10 +1,0 @@
-package duke;
-
-/**
- * Duke Exception class
- */
-public class DukeException extends Exception {
-    public DukeException(String message) {
-        super(message);
-    }
-}
