+++ src/main/java/homie/Deadline.java
@@ -1,4 +1,4 @@
-package duke;
+package homie;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
