+++ src/main/java/homie/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package homie;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -59,7 +59,7 @@ public class DialogBox extends HBox {
         return db;
     }
 
-    public static DialogBox getDukeDialog(String text, Image img) {
+    public static DialogBox getHomieDialog(String text, Image img) {
         var db = new DialogBox(text, img);
         db.flip();
         return db;
