+++ src/main/java/homie/HomieException.java
@@ -1,0 +1,10 @@
+package homie;
+
+/**
+ * Homie Exception class
+ */
+public class HomieException extends Exception {
+    public HomieException(String message) {
+        super(message);
+    }
+}
