+++ src/main/java/homie/Task.java
@@ -1,4 +1,4 @@
-package duke;
+package homie;
 
 import java.io.Serializable;
