+++ src/test/java/homie/TaskListTest.java
@@ -1,4 +1,4 @@
-package duke;
+package homie;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
